Rename to worker warmup and added capability
